This is almost ready
Although the deadline has been extended, I have to move on to other papers, so I will turn it in this afternoon. Please take a look @mariosky @pacastillo, don't ask question or rise issues but directly make your changes on the paper. I will review whatever there is in 6 hours or so and submit it to the conference.